backend/requirements.txt
Django==5.1.2
django-ninja==1.3.0
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.0
psycopg==3.2.3
pyjwt==2.10.1
pytest==8.3.3
pytest-django==4.9.0